Change timeout variables to floating point

+++ packages/engine/tests/experiment/mod.rs
@@ -132,12 +132,12 @@ pub async fn run_test_suite(
         vec![Some(LogLevel::Warning), Some(LogLevel::Trace)]
     };
 
-    let start_timeout = std::env::var("ENGINE_START_TIMEOUT").map_or(10, |val| {
-        val.parse::<u64>()
+    let start_timeout = std::env::var("ENGINE_START_TIMEOUT").map_or(10., |val| {
+        val.parse::<f64>()
             .expect("ENGINE_START_TIMEOUT couldn't be parsed as a u64")
     });
-    let wait_timeout = std::env::var("ENGINE_WAIT_TIMEOUT").map_or(60, |val| {
-        val.parse::<u64>()
+    let wait_timeout = std::env::var("ENGINE_WAIT_TIMEOUT").map_or(60., |val| {
+        val.parse::<f64>()
             .expect("ENGINE_WAIT_TIMEOUT couldn't be parsed as a u64")
     });
